server/requirements.txt
sanic==19.9.0
pytz==2019.3
grpclib==0.4.1
grpcio==1.29.0
grpcio-tools==1.29.0
grpc-proto-validator==0.1.2.dev0
validators==0.18.1
numpy==1.16.0
requests==2.22.0
lxml==4.5.0

Sastrawi==1.0.1
beautifulsoup4===4.8.2
cached-property===1.5.2
bs4~=0.0.1
pymemcache~=3.4.0
pandas~=1.1.3
nltk~=3.5

serialized-redis-interface==0.3.1
redis==3.3.11